+++ snk_snake.h
@@ -13,8 +13,8 @@ typedef struct snk_snake {
     snk_position head_position;
     snk_direction head_direction;
     snk_joint_buffer joints;
-    uint16_t length;
-    uint16_t pending_length;
+    uint32_t length;
+    uint32_t pending_length;
 } snk_snake;
 
 typedef int (*snk_snake_walk_cb)(const snk_position *pos, void *data);
@@ -22,7 +22,7 @@ typedef int (*snk_snake_walk_cb)(const snk_position *pos, void *data);
 int snk_snake_walk(const snk_snake *snake, snk_snake_walk_cb cb, void *cb_data);
 int snk_snake_advance(snk_snake *snake, snk_direction next_direction);
 int snk_snake_get_positions(const snk_snake *snake, size_t *n_positions, snk_position *positions);
-void snk_snake_add_pending_length(snk_snake *snake, uint8_t length);
+void snk_snake_add_pending_length(snk_snake *snake, uint32_t length);
 
 #ifdef __cplusplus
 } /* extern "C" */
